+++ signalfx/event_feed_chart.go
@@ -1,0 +1,158 @@
+package signalfx
+
+import (
+	"encoding/json"
+	"fmt"
+
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func eventFeedChartResource() *schema.Resource {
+	return &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"synced": &schema.Schema{
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     true,
+				Description: "Whether the resource in the provider and SignalFx are identical or not. Used internally for syncing.",
+			},
+			"last_updated": &schema.Schema{
+				Type:        schema.TypeFloat,
+				Computed:    true,
+				Description: "Latest timestamp the resource was updated",
+			},
+			"url": &schema.Schema{
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "URL of the chart",
+			},
+			"name": &schema.Schema{
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Name of the chart",
+			},
+			"description": &schema.Schema{
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Description of the chart (Optional)",
+			},
+			"program_text": &schema.Schema{
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Signalflow program text for the chart. More info at \"https://developers.signalfx.com/docs/signalflow-overview\"",
+			},
+			"viz_options": &schema.Schema{
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Plot-level customization options, associated with a publish statement",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"label": &schema.Schema{
+							Type:        schema.TypeString,
+							Required:    true,
+							Description: "The label used in the publish statement that displays the plot (metric time series data) you want to customize",
+						},
+						"color": &schema.Schema{
+							Type:         schema.TypeString,
+							Optional:     true,
+							Description:  "Color to use",
+							ValidateFunc: validatePerSignalColor,
+						},
+					},
+				},
+			},
+		},
+
+		Create: eventFeedChartCreate,
+		Read:   eventFeedChartRead,
+		Update: eventFeedChartUpdate,
+		Delete: eventFeedChartDelete,
+	}
+}
+
+/*
+  Use Resource object to construct json payload in order to create an event feed chart
+*/
+func getPayloadEventFeedChart(d *schema.ResourceData) ([]byte, error) {
+	payload := map[string]interface{}{
+		"name":        d.Get("name").(string),
+		"description": d.Get("description").(string),
+		"programText": d.Get("program_text").(string),
+	}
+
+	viz := getEventFeedOptions(d)
+	if len(viz) > 0 {
+		payload["options"] = viz
+	}
+
+	return json.Marshal(payload)
+}
+
+func getEventFeedOptions(d *schema.ResourceData) map[string]interface{} {
+	viz := make(map[string]interface{})
+	viz["type"] = "Event"
+
+	return viz
+}
+
+func eventFeedChartCreate(d *schema.ResourceData, meta interface{}) error {
+	config := meta.(*signalfxConfig)
+	payload, err := getPayloadEventFeedChart(d)
+	if err != nil {
+		return fmt.Errorf("Failed creating json payload: %s", err.Error())
+	}
+
+	url, err := buildURL(config.APIURL, CHART_API_PATH, map[string]string{})
+	if err != nil {
+		return fmt.Errorf("[SignalFx] Error constructing API URL: %s", err.Error())
+	}
+
+	err = resourceCreate(url, config.AuthToken, payload, d)
+	if err != nil {
+		return err
+	}
+	// Since things worked, set the URL and move on
+	appURL, err := buildAppURL(config.CustomAppURL, CHART_APP_PATH+d.Id())
+	if err != nil {
+		return err
+	}
+	d.Set("url", appURL)
+	return nil
+}
+
+func eventFeedChartRead(d *schema.ResourceData, meta interface{}) error {
+	config := meta.(*signalfxConfig)
+	path := fmt.Sprintf("%s/%s", CHART_API_PATH, d.Id())
+	url, err := buildURL(config.APIURL, path, map[string]string{})
+	if err != nil {
+		return fmt.Errorf("[SignalFx] Error constructing API URL: %s", err.Error())
+	}
+
+	return resourceRead(url, config.AuthToken, d)
+}
+
+func eventFeedChartUpdate(d *schema.ResourceData, meta interface{}) error {
+	config := meta.(*signalfxConfig)
+	payload, err := getPayloadEventFeedChart(d)
+	if err != nil {
+		return fmt.Errorf("Failed creating json payload: %s", err.Error())
+	}
+	path := fmt.Sprintf("%s/%s", CHART_API_PATH, d.Id())
+	url, err := buildURL(config.APIURL, path, map[string]string{})
+	if err != nil {
+		return fmt.Errorf("[SignalFx] Error constructing API URL: %s", err.Error())
+	}
+
+	return resourceUpdate(url, config.AuthToken, payload, d)
+}
+
+func eventFeedChartDelete(d *schema.ResourceData, meta interface{}) error {
+	config := meta.(*signalfxConfig)
+	path := fmt.Sprintf("%s/%s", CHART_API_PATH, d.Id())
+	url, err := buildURL(config.APIURL, path, map[string]string{})
+	if err != nil {
+		return fmt.Errorf("[SignalFx] Error constructing API URL: %s", err.Error())
+	}
+
+	return resourceDelete(url, config.AuthToken, d)
+}